requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.0
Django==5.1.7
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
packaging==24.1
pillow==11.0.0
pip-review==1.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
